Add tests, fix millions of bugs, refactor a bit
There were bugs left and right, so I fixed them one by one.

Unfortunately, we depend on blackfriday, which is buggy and not commonmark compliant, so it
requires 4 "spaces" in the nesting level to work properly.

Also blackfriday does not parse codeblock in list items and codeblock in blockquote correctly.
So I fail when I see it

---

I still don't really see *why* we need our own markdown formatter just to insert code snippets,
but I don't really understand what pulpMd does exactly, since the code is very hard to read.

If we really need it, let's at least have it working.

---

I also changed it to output spaces instead of tabs, because tabs are really annoying in markdown