render: Support custom formatters
The renderer currently unconditionally reformats Go code samples,
and only Go code samples.

There are a couple issues with this:

- this can lead to inconsistent formatting if some samples
  are fully valid Go code and others aren't
- when desirable, this only works for Go, and nothing else

This proposes a new WithCodeFormatters option for the renderer.
If used, this option replaces the list of code formatters
used by the renderer.

Code formatters are matched by language, optionally with aliases.
By default, only a formatter for "go" (with the alias "Go")
is included -- this matches the prior behavior.

If WithCodeFormatters is called without any arguments,
it disables code formatting completely.