lint: Add go mod tidy
Adds a 'make tidy' and 'make check-tidy' target to 'make lint',
that ensures that go.mod and go.sum are always up-to-date.

Because go mod tidy does not have a dry-run mode that reports
changes it would have made,
this just runs 'go mod tidy' and checks if the file changed.