godoc: All exported entities
A couple exported functions and types were not documented.
This adds some basic godoc to them.

Note that there was a comment on markdownfmt.NewParser
about the possibility of deleting it.
I've deleted that comment because deleting NewParser now
would be a breaking change requiring a v3.
That's not out of the question,
but it seems like a minor thing to cut a v3 over.