main: Don't store exitCode globally
This makes a minor refactor to main
to avoid use of a global exit code variable.

Instead, we declare a mainCmd struct
that stores global program state like exitCode.

This also allows us to change the definition of
stdin, stdout, and stderr,
so that tests can provide varying values.

This contains no logic changes.
All changes are mechanical. Roughly:

- Turn `report(error)` into a method
  that manipulates state on mainCmd.
- Change every function that calls `report` into a method,
  and every function that calls *that* as well.
- Instead of `flag.Parse()`, which operates on `os.Args`,
  call `flag.CommandLine.Parse` with a specific list of arguments.
- Delete `exitCode` variable.