test: Use testify
The tests currently validate assertions manually,
using go-cmp for deep-diffing.
This is nice, and it aligns with
https://github.com/golang/go/wiki/TestComments#assert-libraries
but it results in a massive amount of unnecessary boilerplate.

testify is pretty ubiquitous, and has much nicer error messages.
The failure diffs generated by testify are cleaner than go-cmp.

This switches all tests and assertions to testify.