Move main to cmd/, markdownfmt/ to root
Having a binary named 'markdownfmt' in the root of the project
with a directory also named 'markdownfmt' causes issues.
For one, running `go build` in root results in a failure
because we cannot have a file with the same name as a directory.

Further, it makes for a slightly stuttered library API to have,

    markdownfmt/v3/markdownfmt

This change moves the CLI to cmd/markdownfmt,
and the markdownfmt package to the root of the repository
so that users can import markdownfmt/v3 directly.

Refs #59