Compile with Go 1.17
Go 1.17 is *technically* no longer supported;
upstream Go will not make any fixes to it.

However, as long as supporting it isn't too difficult,
it's not a bad idea to keep the library compiling and running with it.

When we need a feature from 1.18/1.19 that justifies dropping support,
we should feel free to do so.

Refs #69
(Even if that option was added, we would do it only to the v3 version.)